refactor: カレンダーページと現場リスト抽出を共通化してパフォーマンス改善
- 新しい共通カスタムフック useSiteList を作成
- CalendarPage と Sidebar で重複していた現場リスト抽出ロジックを統合
- useMemo による適切なメモ化でパフォーマンス向上
- コードの重複を削減し保守性を改善

Fixes #196
